app/requirements.txt
Flask
requests
Flask-SQLAlchemy
ephem
discord.py
schedule
